# routes/accounts.go
package routes

import (
	"net/http"
	"time"

	"github.com/lakshay35/finlit-backend/models"

	"github.com/gin-gonic/gin"
	accountsService "github.com/lakshay35/finlit-backend/services/account"
	plaidService "github.com/lakshay35/finlit-backend/services/plaid"
	"github.com/lakshay35/finlit-backend/utils/requests"
)

// GetAccountInformation ...
// @Summary Get Account Information
// @Description Gets account information based on access token
// @Tags External Accounts
// @Accept  json
// @Param account body models.Account true "Account payload to get informaion on"
// @Produce  json
// @Security Google AccessToken
// @Success 200 {object} models.PlaidAccount
// @Failure  400 {object} models.Error
// @Router /account/get-account-details [post]
func GetAccountInformation(c *gin.Context) {

	var json models.Account
	err := requests.ParseBody(c, &json)

	if err != nil {
		return
	}

	account, getAccountInformationError := accountsService.GetAccountInformation(json.ExternalAccountID)

	if getAccountInformationError != nil {
		requests.ThrowError(
			c,
			getAccountInformationError.StatusCode,
			getAccountInformationError.Error(),
		)

		return
	}

	c.JSON(http.StatusOK, account)
}

// GetCurrentBalances ...
// @Summary Get Current A/c Balances
// @Description Retrieves live account balances for all accounts attached to an external account registratiom
// @Tags External Accounts
// @Accept  json
// @Produce  json
// @Security Google AccessToken
// @Success 200 {array} models.PlaidGetBalancesResponse
// @Failure  400 {object} models.Error
// @Router /account/live-balances [get]
func GetCurrentBalances(c *gin.Context) {
	response, err := plaidService.PlaidClient().GetBalances("accessToken")

	if err != nil {
		requests.ThrowError(
			c,
			http.StatusBadRequest,
			err.Error(),
		)
	}

	c.JSON(http.StatusOK, response.Accounts)
}

// GetTransactions ...
// @Summary Get Transactions
// @Description Gets all transactions for the  past 30 days
// @Tags External Accounts
// @Accept  json
// @Produce  json
// @Param body body models.Account true "Account payload to identify transactions with"
// @Security Google AccessToken
// @Success 200 {array} models.PlaidTransaction
// @Failure 403 {object} models.Error
// @Router /account/transactions [post]
func GetTransactions(c *gin.Context) {
	var json models.Account
	err := requests.ParseBody(
		c,
		&json,
	)

	if err != nil {
		return
	}

	transactions, transactionsError := accountsService.GetTransactions(
		json.ExternalAccountID,
		time.Now().Local().Add(-30*24*time.Hour).Format("2006-01-02"),
		time.Now().Local().Format("2006-01-02"),
	)

	if transactionsError != nil {
		requests.ThrowError(
			c,
			transactionsError.StatusCode,
			transactionsError.Message,
		)

		return
	}

	c.JSON(http.StatusOK, transactions)
}

// CreateLinkToken ...
// Creates link token
// @Summary Create Link Token
// @Description Creates a link token to setup UI for generating public tokens
// @Tags External Accounts
// @Accept  json
// @Produce  json
// @Security Google AccessToken
// @Success 200 {object} models.LinkTokenPayload
// @Failure 400 {object} models.Error
// @Router /account/create-link-token [get]
func CreateLinkToken(c *gin.Context) {
	linkToken, err := accountsService.LinkTokenCreate(nil)

	if err != nil {
		requests.ThrowError(
			c,
			err.StatusCode,
			err.Error(),
		)
		return
	}

	c.JSON(http.StatusOK, models.LinkTokenPayload{
		LinkToken: linkToken,
	})
}

type httpError struct {
	errorCode int
	error     string
}

func (httpError *httpError) Error() string {
	return httpError.error
}

// GetAllAccounts ...
// @Summary Get all registered external accounts
// @Description Gets a list of all external accounts registered via Plaid
// @Tags External Accounts
// @Accept  json
// @Produce  json
// @Security Google AccessToken
// @Success 200 {array} models.Account
// @Failure 403 {object} models.Error
// @Failure 404 {object} models.Error
// @Router /account/get [get]
func GetAllAccounts(c *gin.Context) {
	user := requests.GetUserFromContext(c)
	accounts, err := accountsService.GetAllExternalAccounts(user.UserID)

	if err != nil {
		requests.ThrowError(
			c,
			err.StatusCode,
			err.Error(),
		)
		return
	}

	c.JSON(http.StatusOK, accounts)
}

// RegisterAccessToken ...
// @Summary Register Access Token
// @Description Creates a permanent access token based on public token
// @Tags External Accounts
// @Accept  json
// @Param body body models.PublicTokenPayload true "Token Payload for registering access token"
// @Produce  json
// @Security Google AccessToken
// @Success 201
// @Failure 403 {object} models.Error
// @Router /account/register-token [post]
func RegisterAccessToken(c *gin.Context) {
	var json models.PublicTokenPayload
	err := c.BindJSON(&json)

	if err != nil {
		requests.ThrowError(
			c,
			http.StatusBadRequest,
			err.Error(),
		)
		return
	}

	user := requests.GetUserFromContext(c)

	accessTokenError := accountsService.RegisterAccessToken(json.Token, user.UserID)

	if accessTokenError != nil {
		requests.ThrowError(
			c,
			accessTokenError.StatusCode,
			accessTokenError.Error(),
		)
		return
	}

	c.JSON(http.StatusCreated, nil)
}